requirements.txt
streamlit==1.37.1
ultralytics==8.2.103
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
gdown==5.1.0
torch>=2.0.0
torchvision>=0.15
